requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler
gunicorn
python-dotenv
Flask-WTF
email-validator